+++ src/actions/workspaces/option/projects/get_options.ts
@@ -1,5 +1,4 @@
-import { getWorkspaceProjects } from "../../../../utils/api";
-import { getAPIErrorMessage } from "../../../../utils/helper";
+import retrieveProjects from "../../../../common/retrieveProjects";
 import { Project, Workspace } from "../../../../utils/interfaces";
 
 // @description Get all workspace projects
@@ -8,27 +7,12 @@ const returnOptions = () => {
 	let projects: Project[] = [];
 	if (options.workspaces) {
 		const workspace = options.workspaces as Workspace;
-		const projectsResponse = getWorkspaceProjects(workspace.gid);
-		if (projectsResponse.status === 200) {
-			projects = projectsResponse.response.data as Project[];
-			projects = projects.map(project => ({
-				...project,
-				description: getProjectDescription(project)
-			}));
-		} else {
-			log(getAPIErrorMessage(projectsResponse.response), "#FF5733");
-		}
+		projects = retrieveProjects("workspace", workspace.gid);
 	}
 
 	return {
 		add: projects
 	};
-}
-
-const getProjectDescription = (project: Project) => {
-	let description = project.notes ? project.notes : project.name;
-	description += ` ${project.archived ? "(Archived)" : ""}`;
-	return description;
 }
 
 export default returnOptions;
