+++ src/common/retrieveProjects.ts
@@ -1,0 +1,35 @@
+import { getTeamProjects, getWorkspaceProjects } from "../utils/api";
+import { getAPIErrorMessage } from "../utils/helper";
+import { DecodedAPIResponse, Project } from "../utils/interfaces"
+
+// gid can be workspaceid or teamid
+
+const retrieveProjects = (type: "workspace" | "team", gid: number): Project[] => {
+	let projects: Project[] = [];
+	let projectsResponse = <DecodedAPIResponse>{};
+	if (type === "workspace") {
+		projectsResponse = getWorkspaceProjects(gid);
+	}
+	if (type === "team") {
+		projectsResponse = getTeamProjects(gid);
+	}
+	if (projectsResponse.status === 200) {
+		projects = projectsResponse.response.data as Project[];
+		projects = projects.map(project => ({
+			...project,
+			description: getProjectDescription(project)
+		}));
+	} else {
+		log(getAPIErrorMessage(projectsResponse.response), "#FF5733");
+	}
+	return projects;
+}
+
+const getProjectDescription = (project: Project) => {
+	let description = project.notes ? project.notes : project.name;
+	description += ` ${project.archived ? "(Archived)" : ""}`;
+	return description;
+}
+
+
+export default retrieveProjects;
